user-edit-update

+++ db/migrate/20191220070043_create_foods.rb
@@ -3,11 +3,11 @@ class CreateFoods < ActiveRecord::Migration[5.0]
     create_table :foods do |t|
       t.string :name, null: false
       t.integer :calorie
-      t.integer :carbohydrate
-      t.integer :protein
-      t.integer :fat
-      t.integer :salt_equivalent
-      t.integer :dietary_fiber
+      t.float :carbohydrate
+      t.float :protein
+      t.float :fat
+      t.float :salt_equivalent
+      t.float :dietary_fiber
       t.timestamps
     end
   end
